Move base to cli, utils to base
