+++ unidist/cli/base/runner.py
@@ -8,7 +8,7 @@ import os
 import subprocess
 
 
-from unidist.cli.backends.utils import Defaults
+from unidist.cli.base.utils import Defaults
 
 
 class BackendRunner:
